services/market_data/requirements.txt
# FastAPI and ASGI server
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.7.1
asyncpg==0.24.0

# Redis
redis==3.5.3
aioredis==2.0.0

# Data processing
pandas==1.3.3
numpy==1.21.2
yfinance==0.1.63
pytz==2021.1

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0

# Monitoring and logging
prometheus-client==0.11.0
python-json-logger==2.0.2

# Utilities
python-dotenv==0.19.0
pydantic==1.8.2
email-validator==1.1.3
tenacity==8.0.1 